Conflicts:
	samples/Cpp/TestCpp/Classes/ExtensionsTest/CocoStudioArmatureTest/ArmatureScene.h

+++ cocos/2d/CCEventDispatcher.h
@@ -136,12 +136,12 @@ private:
         
         inline std::vector<EventListener*>* getFixedPriorityListeners() const { return _fixedListeners; };
         inline std::vector<EventListener*>* getSceneGraphPriorityListeners() const { return _sceneGraphListeners; };
-        inline long getGt0Index() const { return _gt0Index; };
-        inline void setGt0Index(long index) { _gt0Index = index; };
+        inline int getGt0Index() const { return _gt0Index; };
+        inline void setGt0Index(int index) { _gt0Index = index; };
     private:
         std::vector<EventListener*>* _fixedListeners;
         std::vector<EventListener*>* _sceneGraphListeners;
-        long _gt0Index;
+        int _gt0Index;
     };
     
     /** Adds event listener with item */
